start over

+++ robot/script/neural_net.py
@@ -85,7 +85,7 @@ class NeuralNet(object):
 if __name__ == '__main__':
     try:
         train_iter = int(sys.argv[2])
-        threshold = int(sys.argv[4])
+        threshold = float(sys.argv[4])
         nn = NeuralNet(train_iter, threshold)
     except rospy.ROSInterruptException:
         pass
